Create README.md
Project Description
As a Data Analyst who works for the education tech company Udemy, I was asked by my manager, the Head of Curriculum, to analyze the data on course revenue. I was provided with data on courses from different topics to understand where opportunities to increase revenue could lie, and track the performance of courses.
My manager suggested encouraging Web Development courses to have an increase in charge because she strongly believed that these are the most popular courses. She needed to send a report to the CEO in three weeks on how they would increase the next quarterly earnings.
I was provided with four different spreadsheets which required consolidation into one spreadsheet, and the creation of pivot tables and charts for the analysis.
From the analysis, I was able to break down the opportunities available to increase revenue in the courses that had such potentials.

Overview
Udemy, Inc. is a for-profit massive open online course provider aimed at professional adults and students. They have over 183000 courses and 40 million students. They offer a wide range of courses from programming, data science, and lots more. Each course has numerous subjects offered at the beginner, intermediate and expert levels either for free or paid.

Background of the project
Udemy is a for-profit company, and revenue generation is an integral part of its business. A performance review on all the subjects and their courses was required. The aim was to understand where opportunities existed for an increase in revenue in those subjects with emphasis on the Web Development since it was assumed to be the most popular subject on the platform.

Framing the Problem
The project duration was for 2 weeks because the report was to reach the CEO in 3 weeks. The project started by getting an understanding of the situation and gathering the right data. After collecting the right data, I prepared for the analysis process.

Design 
This project required processing the datasets from the scratch. I was provided with four separate datasets. I used Google sheets to consolidate all the files into one by importing the first file as “replace spreadsheet” and the others as “Append rows to current sheet”. 
I cleaned the dataset by removing duplicates, blank cells, etc. I then created new columns using spreadsheet functions and formulas. I added some new columns to the table using the RIGHT/LEFT Functions, IF function, and LOOKUP after which I created pivot tables and charts for the analysis. Finally, I used the VLOOKUP function to determine the top 20 most subscribed courses.

Analysis
I made use the following metrics to create the Pivot tables and charts:
•	Total number of subscribers for each subject 
•	Average number of subscribers for each subject 
•	Average cost per subject at each level 
•	Average content duration for each subject 
•	Average rating per subject for each level 
•	Average number  of lectures per subject and level


Analysis Explained
Employing the Root Cause Analysis Process, I asked some questions while trying to gain insights to the data.
Recall that from the background of the problem, Web Design was assumed to be the most popular subject on the platform. With this information, I dug deeper to find out what made web Development so well liked.
The first question I asked was:
1. Why is Web Development assumed to be the most popular subject on the platform?
To answer this question, I used pivot table to have an idea of the number of subscribers per subject and came up with the following insights.
Shockingly, that initial assumption was wrong. As seen from the result, Business Finance had the highest number of subscribers with 10,491,014 while Web Development had 80,295 subscribers which was the least with a very wide margin.
This result made me to be attentive to the subject with the largest number subscribers – Business Finance. Subsequent questions involved the two subjects.

2. Why does the Web Development subject have the least number of subscribers?
This question required an in-depth look into the dataset to discover the various factors that might be responsible for this subject having a tremendously low amount of learners. I looked into one of the factors which was the subject rating. The analysis showed clearly that the Web Development subject had the lowest rating. This discovery led to the next question.

3. Why does the Web Development subject have the lowest rating compared to the other three subjects?
Further analysis showed that Web Development had highest price, longest content duration, and the most number of lectures compared to the other subjects. These were the factors that were keeping subscribers from enrolling into the Web Development subject. 
At this point, I had gotten to the root cause of the problem but still needed to explore the other subjects with particular attention to the course with the highest number of subscribers – Business Finance.

4. Why is Business Finance the most popular of the four subjects?
The wide gap between Business Finance subscribers and the subject with the least subscribers was a call for an extended analysis of the dataset. To answer the question, I revisited the dataset to compare the price, rating, content duration, and number of lectures. The analysis showed that Business Finance had the second highest rating, lowest price, shortest content duration, and the least number of lectures. This result led to the next question.

5. Why did the Udemy coordinators fix a very high price, make long contents and have many lectures for the Web Development subject as opposed to that of Business Finance? What informed that decision?
It was discovered from the Head of curriculum at Udemy that because the Web Development subject had so many intricacies, it required more course content and lectures to break it down to its simplest form so that the subscribers could grasp it better. This also meant attaching the highest price to it.

Recommendations
1.	An immediate downward review of the price of the Web Development subject should be implemented while content duration and number of lectures for the same subject can be also reduced with time due to its intricacies.
2.	Due to the popularity of the Business Finance subject, it is advised that the price should be increased by a little percentage while continually monitoring the number of subscribers.
3.	Graphic Design subject had the highest rating compared to the other subjects. This means that the subscribers benefit greatly from the course content. Another opportunity for a price increase lies here.
4.	The number of lectures for two subjects that have the least number of subscribers should be reduced because the trend from the analysis shows that the more lectures a subject has, the fewer subscribers it would have. The subjects are Musical Instruments and Web Development.

Conclusion
A myriad of opportunities exist for revenue generation for Udemy in the next quarter. The only thing required is to try to adhere to the recommendations above while I continually monitor the performance for each subject and their courses.


Note: This data is fictional. It was one of the projects I worked on while undergoing my data analysis course on the Entry Level Platform.
 The visualization tool that I used was Tableau and Google sheet. I used Tableau because it was easy for me to work with considering the large dataset that resulted after the consolidation. Google sheet made it easier for me to quickly analyze the data to have a quick overview using the pivot table and spreadsheet functions.